+++ changelog.md
@@ -1,6 +1,7 @@
 # To Do
 
-- [ ] Move Edit and Record Payment modals to their own component classes
+- [x] Move Edit and Record Payment modals to their own component classes
+
 - [ ] Add API integrations (details TBD)
 
 # Changelog
@@ -8,3 +9,8 @@
 ## 2025-08-27
 
 - [x] added Edit and Record Payment modals
+
+## 2025-08-28
+
+- [x] added new app icon
+- [x] added progressive web application